Moving classes to different files to be compliant with StyleCop analyzers
